### 2019-05-09 77108ed
added missing .bib tag in paper.md
