### 2019-05-09 #1438 @labarba
@abhishektha — Please see the [example paper](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography) in the documentation, and fix your paper—it's missing the bibliography entry in the metadata.

Also, I know that you suggested @Kevin-Mattheus-Moerman as handling editor, and this is in his area of interest, but he has a high editorial load right now. Unless he pops in here and says he's willing to handle one more paper, I'll look for another editor.
