### 2019-05-09 #1438 @abhishektha
@labarba, I've updated the paper and uploaded it to the main branch of the repository. Does this need to be tagged as a release to be recompiled?

And thank you for searching for another editor!
