### 2019-05-09 #1438 @labarba
You're missing a comma after `@article{Windolf:2008`
