+++ paper.bib
@@ -13,7 +13,7 @@ keywords = "Whole-body-scanners, Anthropometry, Garments",
 abstract = "An overview of whole body scanners in 1998 (H.A.M. Daanen, G.J. Van De Water. Whole body scanners, Displays 19 (1998) 111–120) shortly after they emerged to the market revealed that the systems were bulky, slow, expensive and low in resolution. This update shows that new developments in sensing and processing technology, in particular in structured light scanners, have produced a new generation of easy to transport, fast, inexpensive, accurate and high resolution scanners. The systems are now moving to the consumer market with high impact for the garment industry. Since the internet sales of garments is rapidly increasing, information on body dimensions become essential to guarantee a good fit, and 3D scanners are expected to play a major role."
 }
 
-@article{Windolf:2008
+@article{Windolf:2008,
 author={Windolf,Markus and Götzen,Nils and Morlock,Michael},
 year={2008},
 title={Systematic accuracy and precision analysis of video motion capturing systems--exemplified on the Vicon-460 system},
