### 2019-05-09 #1438 @abhishektha
@labarba looks like I finally got the pdf compiled, thank you for your help so far!
