### 2019-05-10 #1438 @labarba
@moorepants — Do you think you could contribute a review for this JOSS submission? Title: "DynaMo: Dynamic Body Shape and Motion Capture with Intel RealSense Cameras"
Let me know!
