### 2019-05-10 #1438 @labarba
@melund — Might you be able to contribute a review of this submission for JOSS? Title: "DynaMo: Dynamic Body Shape and Motion Capture with Intel RealSense Cameras"
