### 2019-05-10 #1438 @melund
@labarba Yes. I would like that. But I don't have access to the cameras, so there is certain parts I won't be able to review in detail. If that is ok then I am in :)
