### 2019-05-19 #1438 @labarba
👋  @ixjlyons — would you be willing to contribute a review for JOSS on this submission? "DynaMo: Dynamic Body Shape and Motion Capture with Intel RealSense Cameras"
