### 2019-05-21 #1466 @whedon
[REVIEW]: DynaMo: Dynamic Body Shape and Motion Capture with Intel RealSense Cameras
**Submitting author:** @abhishektha (<a href="http://orcid.org/0000-0003-4260-0488">Abhishektha Boppana</a>)
**Repository:** <a href="https://github.com/anderson-cu-bioastronautics/dynamo_realsense-capture" target ="_blank">https://github.com/anderson-cu-bioastronautics/dynamo_realsense-capture</a>
**Version:** v1.1
