### 2019-06-07 716489b
Modified environment.yml to remove specific builds, should help cross platform installation. Edited directories for testscript.py to allow for correct finding of sample files.
