+++ environment.yml
@@ -1,64 +1,29 @@
-name: realsense
+name: dynamo
 channels:
   - sirokujira
   - intel
   - conda-forge
-  - r
+  - anaconda
   - defaults
 dependencies:
-  - boost=1.64.0=py36_vc14_4
-  - boost-cpp=1.64.0=vc14_1
-  - eigen=3.3.3=0
-  - filelock=3.0.10=py_0
-  - icu=58.2=vc14_0
-  - jpeg=9c=hfa6e2cd_1001
-  - libtiff=4.0.10=h36446d0_1001
-  - libwebp=1.0.2=hfa6e2cd_1
-  - opencv=4.0.1=py36hb76ac4c_200
-  - qt=5.6.2=vc14_1
-  - bzip2=1.0.6=vc14_17
-  - certifi=2018.1.18=py36_2
-  - icc_rt=2019.3=intel_203
-  - intel-openmp=2019.3=intel_203
-  - intelpython=2019.3=0
-  - libpng=1.6.36=vc14_2
-  - mkl=2019.3=intel_203
-  - mkl_fft=1.0.11=py36h7b7c402_0
-  - mkl_random=1.0.2=py36h7b7c402_4
-  - nose=1.3.7=py36_18
-  - numpy=1.16.1=py36h7b7c402_3
-  - numpy-base=1.16.1=py36_3
-  - openssl=1.0.2r=vc14_0
-  - pip=10.0.1=py36_0
-  - python=3.6.5=11
-  - setuptools=39.0.1=py36_0
-  - sqlite=3.23.1=vc14_1
-  - tbb=2019.4=vc14_intel_203
-  - tbb4py=2019.4=py36_intel_0
-  - tcl=8.6.4=vc14_22
-  - tk=8.6.4=vc14_28
-  - vc=14.0=2
-  - vs2015_runtime=14.0.25420=intel_2
-  - wheel=0.31.0=py36_3
-  - wincertstore=0.2=py36_3
-  - xz=5.2.3=vc14_2
-  - zlib=1.2.11=vc14h21ff451_5
-  - flann=1.9.1=h0aa8a5b_0
-  - pcl=1.8.1=h0b3e05b_0
-  - python-pcl=0.3.0=py36h0881630_0
-  - qhull=2015.2=hd783fd8_0
-  - vtk=8.0.1=h91cdee5_0
-  - pip:
-    - cloudpickle==0.8.0
-    - configparser==3.7.3
-    - cython==0.29.6
-    - dask==1.1.3
-    - gias2==0.6.6
-    - networkx==2.2
-    - pillow==5.4.1
-    - pyrealsense2==2.19.1.580
-    - pywavelets==1.0.2
-    - scikit-image==0.14.2
-    - toolz==0.9.0
-prefix: C:\Users\abhis\Anaconda3\envs\realsense
+  - intel-openmp=2019.3
+  - intelpython=2019.3
+  - jpeg=9c
+  - mkl=2019.3
+  - mkl_fft=1.0.11
+  - mkl_random=1.0.2
+  - numpy=1.16.1
+  - numpy-base=1.16.1
+  - opencv=4.0.1
+  - pcl=1.8.1
+  - pip=10.0.1
+  - python=3.6.5
+  - python-pcl=0.3
+  - scikit-learn=0.20.0
+  - scipy=1.2.1
+  - vtk=8.0.1
 
+  - pip:
+    - pptk==0.1.0
+    - pyrealsense2==2.22.0.864
+prefix: C:\Users\AndersonBioastroTest\Anaconda3\envs\dynamo
