### 2019-06-10 b3f6883
added community guidelines and additional Anaconda setup instructions
