### 2019-06-11 #1466 @ixjlyons
Apologies for the delay. Here is my review:

### Paper

The paper is written clearly and offers insight into the functionality of the library. A few minor points:

- Unresolved reference in first paragraph.
- Figure 1 not showing.
- 2nd from last paragraph: "synchroniously" -> "synchronously"
- I think the last paragraph describing some research applications could be better integrated into the background section of the paper.

### Docs

The documentation is distributed into the README, a couple other markdown documents, and a Jupyter notebook. These work together to cover much of the library's functionality and usage, but you could benefit from a dedicated documentation site with API documentation generated from the docstrings. Every function appears to have docstrings and those files are browsable through GitHub, so I will defer to @labarba on whether or not that satisfies the functionality documentation check.

A few other minor issues:

- Links to files are broken in this section: https://github.com/anderson-cu-bioastronautics/dynamo_realsense-capture/tree/b3f688389712a5be60f373a3fbefa45542519f01#included-functions
- You may want to mention that the `kabsch` function in the tutorial notebook is a function that's available from the library and not something the user has to define.
- The image at the bottom of the tutorial notebook is missing.

### Code/Functionality

Unfortunately, I wasn't able to do much with Dynamo being stuck with Linux (no pyrealsense2) and lacking D4XX cameras, so I'm leaving the installation, functionality, and performance checks un-checked.

I notice that some of the modules (e.g. `calculate_rmsd.py`) can be run as scripts. Consider packaging that functionality as [console scripts](https://python-packaging.readthedocs.io/en/latest/command-line-scripts.html#the-console-scripts-entry-point), or perhaps mention that they can be run this way in the documentation.

Just out of curiosity, have you considered additional storage formats aside from pickle? HDF5 or other formats might work well and could provide better portability.

### Other

The tests are not automated, but as far as I know the instructions for running the tests manually suffices for the "automated tests" check. This library poses some challenges for automated testing (with dependence on hardware), though some of the computational aspects of the software *could* be automatically tested.

### Overall

Overall I think Dynamo would benefit from a dedicated documentation page with API documentation and perhaps automated testing of functionality that doesn't depend on presence of specific hardware. The documentation seems fairly complete, but a more cohesive presentation of it could help newcomers to the library.
