### 2019-06-16 #1466 @melund
@labarba. I am also done with my review. Specific feedback has been handled on the [issue tracker:](https://github.com/anderson-cu-bioastronautics/dynamo_realsense-capture/issues)

Since I haven't got access to the cameras I couldn't review all part of the functionality. But what I could check and review does comply with the requirements of JOSS.

I agree with @ixjlyons about the lack of API documentation, but it is not blocking in my view. I think the most important comment is the [lack of a conda-forge package](#2). It would help guarantee that this package continues to work in the future. However, the current state is in accordance with the requirements of JOSS.
