### 2019-06-19 #1466 @abhishektha
Hi all,

Thank you for the insightful comments and help with getting DynaMo up to JOSS standards so far. I've fixed the errors in the paper and tutorial Jupyter Notebook.

Is there anything else needed from me? I am a little unclear if I need to add to the documentation to proceed with the paper. Would a new section on the github which shows how to call each function of the package meet this requirement? If so I'm happy to get that done soon.

I will also work on getting this package onto conda-forge. It took me a while to catch up on reading how conda-forge recipes work but I think once we have pyrealsense2 on conda-forge, I can easily create a package for DynaMo.

Thanks again for all you contributions so far!
