### 2019-06-20 #1466 @melund
> It took me a while to catch up on reading how conda-forge recipes work but I think once we have pyrealsense2 on conda-forge, I can easily create a package for DynaMo.

That is great, but I think you will have to initiate the work with a `pyrealsense2` conda-forge package. I think a win only package would be Ok for the first iteration. If you start the work, there is a good chance others will chip in and help.
