### 2019-06-25 058d68d
added method for multicamera calibration in calibration.newIterative()
