### 2019-06-26 bbf2199
fixed missing references in marker identification
