### 2019-07-10 #1466 @abhishektha
@labarba Is there anything specific that I need to do for us to proceed with the review? I believe I addressed all the reviewer's comments (please let me know if I didn't). I am a bit confused from the reviews if I need to improve the documentation to meet the JOSS standards (since @melund mentioned it was non-blocking), so please let me know what is needed from me.
