### 2019-07-15 #1466 @labarba
Hi @abhishektha — neither of the two reviewers have been able to check the functionality of the software, due to the hardware dependency (cameras) and lack of access. Accepting the paper under those conditions is awkward enough, so I would request that you make every effort at improvements that you can, given the reviewer comments.

Since you already have docstrings throughout, it should not be too hard to deploy API documentation using Sphinx? Please have a look at that and let me know what you think.

In the meantime, would you be able to suggest someone who may have access to the cameras to provide a check of functionality through a partial & supplementary review?
