### 2019-07-16 #1466 @abhishektha
Thanks for the reply @labarba! I can definitely deploy API documentation through Sphinx, and will let you and the reviewers know when that is done.

Would it be okay if we had someone independently review functionality of the code using our own camera setup in our lab? We have a few colleagues in the field who have not worked with us on this project but may be able to provide a check of functionality. We would simply provide them access to our set of cameras and allow them to hook up their own computer to check functionality.
