### 2019-07-22 06ad823
added api documentation under documentation folder and added to readme.md
