+++ documentation/api.md
@@ -1,0 +1,398 @@
+# dynamo
+
+# dynamo.calibration
+
+## load
+```python
+load(fileName)
+```
+
+Calibration parameters for previously connected cameras are loaded from a pickle file format.
+Calibration parameters for each camera include a 4x4 transformation matrix and rmsd error of calibration
+
+Parameters
+----------
+fileName : str
+    Filename of stored calibration parameters.
+
+Returns
+-------
+devicesTransformation : dict
+    Keys of camera's serial number holding dictionary of calibration parameters per camera
+
+Example
+-----
+load('savedCalibration.cal')
+
+## new
+```python
+new(fileName, deviceManager, chessboardHeight, chessboardWidth, chessboardSquareSize)
+```
+
+New calibration parameters for each connected camera and are created and saved in a pickle file format.
+Calibration parameters for each camera include a 4x4 transformation matrix and rmsd error of calibration
+
+Parameters
+----------
+fileName : str
+    Filename to store calibration parameters.
+
+deviceManager : DeviceManager object
+    realsense_device_manager object which manages connections to all cameras
+
+chessboardHeight : int
+    Number of chessboard intersections defining height of target chessboard
+
+chessboardWidth : int
+    Number of chessboard intersections defining width of target chessboard
+
+chessboardSquareSize : float
+    Dimension of side of chessboard (m)
+
+Returns
+-------
+devicesTransformation : dict
+    dictionary with keys of camera's serial number holding dictionary of calibration parameters per camera
+
+Example
+-----
+    new('savedCalibration.cal')
+
+## newIterative
+```python
+newIterative(fileName, deviceManager, cameraList, chessboardHeight, chessboardWidth, chessboardSquareSize)
+```
+
+New calibration parameters for each connected camera and are created and saved in a pickle file format.
+Calibration parameters for each camera include a 4x4 transformation matrix and rmsd error of calibration
+
+Parameters
+----------
+fileName : str
+    Filename to store calibration parameters.
+
+deviceManager : DeviceManager object
+    realsense_device_manager object which manages connections to all cameras
+
+cameraList : list
+    list of serial numbers to calibrate cameras in order
+
+chessboardHeight : int
+    Number of chessboard intersections defining height of target chessboard
+
+chessboardWidth : int
+    Number of chessboard intersections defining width of target chessboard
+
+chessboardSquareSize : float
+    Dimension of side of chessboard (m)
+
+Returns
+-------
+deviceTransformations : dict
+    dictionary with keys of camera's serial number holding dictionary of calibration parameters per camera
+
+Example
+-----
+    new('savedCalibration.cal')
+
+## detectChessboard
+```python
+detectChessboard(deviceManager, cameraSet, chessboardHeight, chessboardWidth, chessboardSquareSize)
+```
+
+Chessboard locations are computed for each connected RealSense camera
+
+Parameters
+----------
+deviceManager : DeviceManager object
+    realsense_device_manager object which manages connections to all cameras
+
+cameraSet : list
+    list of camera serial numbers to detect the chessboard
+
+chessboardHeight: int
+    Number of chessboard intersections defining height of target chessboard
+
+chessboardWidth: int
+    Number of chessboard intersections defining width of target chessboard
+
+chessboardSquareSize: float
+    Dimension of side of chessboard (m)
+
+Returns
+-------
+chessboardLocations : dict
+    dictionary with keys of camera's serial number holding detected corners, local 2D and 3D coordinates of detected corners, and their valid depth points
+
+Example
+-----
+    detectChessboard(deviceManager, chessboardHeight, chessboardWidth, chessboardSquareSize)
+
+## poseTransformation
+```python
[…]
