### 2019-08-06 d7c8b92
updated tooltips for API documentation, added API documentation
