+++ dynamo/realsense_device_manager.py
@@ -5,12 +5,31 @@
 ## https://github.com/IntelRealSense/librealsense/blob/master/wrappers/python/examples/box_dimensioner_multicam/realsense_device_manager.py ##
 ## Modified functions are denoted with *Modified* and new functions are denoted with *New*                                                  ##
 ##############################################################################################################################################
+__doc__ = \
+"""
+Class for managing Intel RealSense Devices
+This code has been appended to and modified from the librealsense box_dimensioner_multicam example:                                      
+https://github.com/IntelRealSense/librealsense/blob/master/wrappers/python/examples/box_dimensioner_multicam/realsense_device_manager.py 
 
+Modified functions are denoted with *Modified* and new functions are denoted with *New*               
+
+Distributed as a module of DynaMo: https://github.com/anderson-cu-bioastronautics/dynamo_realsense-capture
+"""
 import pyrealsense2 as rs
 
 
 class Device:
     def __init__(self, pipeline, pipeline_profile):
+        """
+        Class to manage each Intel RealSense D4XX Device
+
+        Parameters
+        ----------
+        pipeline : rs.pipeline() object
+
+        pipeline_profile : enabled rs.pipeline() object
+
+        """
         self.pipeline = pipeline
         self.pipeline_profile = pipeline_profile
 
@@ -19,12 +38,12 @@ def enumerate_connected_devices(context):
     """
     Enumerate the connected Intel RealSense devices
 
-    Parameters:
+    Parameters
     -----------
-    context 	 	  : rs.context()
+    context : rs.context()
         The context created for using the realsense library
 
-    Return:
+    Return
     -----------
     connect_device : array
         Array of enumerated devices which are connected to the PC
@@ -37,27 +56,32 @@ def enumerate_connected_devices(context):
     return connect_device
 
 
-def post_process_depth_frame(depth_frame, decimation_magnitude=1.0, spatial_magnitude=2.0, spatial_smooth_alpha=0.5,
-    spatial_smooth_delta=20, temporal_smooth_alpha=0.4, temporal_smooth_delta=20):
+def post_process_depth_frame(depth_frame, decimation_magnitude=1.0, spatial_magnitude=2.0, spatial_smooth_alpha=0.5, spatial_smooth_delta=20, temporal_smooth_alpha=0.4, temporal_smooth_delta=20):
     """
     Filter the depth frame acquired using the Intel RealSense device
 
-    Parameters:
+    Parameters
     -----------
-    depth_frame 	 	 	 : rs.frame()
+    depth_frame : rs.frame()
         The depth frame to be post-processed
+
     decimation_magnitude : double
         The magnitude of the decimation filter
-    spatial_magnitude 	 : double
-                            The magnitude of the spatial filter
-    spatial_smooth_alpha	 : double
-                            The alpha value for spatial filter based smoothening
-    spatial_smooth_delta	 : double
-                            The delta value for spatial filter based smoothening
+
+    spatial_magnitude : double
+        The magnitude of the spatial filter
+
+    spatial_smooth_alpha : double
+        The alpha value for spatial filter based smoothening
+
+    spatial_smooth_delta : double
+        The delta value for spatial filter based smoothening
+
     temporal_smooth_alpha : double
-                            The alpha value for temporal filter based smoothening
+        The alpha value for temporal filter based smoothening
+
     temporal_smooth_delta : double
-                            The delta value for temporal filter based smoothening
+        The delta value for temporal filter based smoothening
 
 
     Return:
@@ -98,11 +122,11 @@ def post_process_depth_frame(depth_frame, decimation_magnitude=1.0, spatial_magn
 class DeviceManager:
     def __init__(self, context, pipeline_configuration):
         """
-        Class to manage the Intel RealSense devices
+        Class to manage all connected Intel RealSense devices
 
[…]
