### 2019-08-06 #1466 @abhishektha
@labarba Sorry for the delay, I had some trouble using Sphinx with the numpydoc style of tooltips I had been using, but its all fixed now and we finally have API documentation hosted on Github Pages and linked from the readme. @ixjlyons, @melund, please let me know if this new API documentation meets the standards.

I would like to suggest @alcantarar to contribute functionality checks to the review. He has access to our cameras but has not worked on the development of this project.
