### 2019-08-07 #1466 @labarba
👋  @alcantarar — would you be willing to contribute a partial review of this JOSS submission, given that the assigned reviewers both lack access to the needed hardware to confirm functionality?
