### 2019-08-07 #1466 @alcantarar
I would be willing to assist with the review.
Should I just open issues in the Dynamo repo as needed and comment below when my review is complete? I was going to follow the functionality section of the review checklist you've provided above.
