### 2019-08-07 #1466 @labarba
Hi @alcantarar — thanks for agreeing to provide an additional review. I have added a review checklist for you. If you don't want to check off all items, that is OK, but we do need the COI and CoC responses from you, and especially need the functionality items that the other two reviewers could not check. Godspeed!
