### 2019-08-13 3148ee9
self-populated camera serial numbers in calibration.new, removed autoformat of pcl in dynamo.view
