### 2019-08-14 anderson-cu-bioastronautics/dynamo_realsense-capture#6 @alcantarar
stream.py function documentation out of date
[Example Tutorial documentation](https://github.com/anderson-cu-bioastronautics/dynamo_realsense-capture/blob/dynamo-release/examples/examples.md#camera-streaming-to-disk) gives this example:
`stream.start(deviceManager, transformationMatrices, 'saveFolder', iteration, time)
`

but `stream.start()` only takes in 4 arguments:
`def start(deviceManager, deviceTransformations, saveDirectory, stream_time):
`

Documentation in `stream.start()` matches Example Tutorial documentation, but not actual function design. I suggest going through function documentation and checking that number of arguments and argument names match. Outdated documentation persists in `stream.processThread()` and  `stream.captureThread()`, for example.

 Part of https://github.com/openjournals/joss-reviews/issues/1466
