### 2019-08-14 b972b86
Removed hard coding of camera serial for frame number and outdated iteration argument for stream.start()
