### 2019-08-14 anderson-cu-bioastronautics/dynamo_realsense-capture#6 @abhishektha
I've removed the outdated iteration argument from stream.start() and its references in all the documentation. This should now be updated in both the example tutorial documentation and the API documentation.
