### 2019-08-14 0c71c0c
missing return parameter in calibration.load()
