### 2019-08-14 8a29353
updated saving of calibration objects to ensure file was handled properly
