### 2019-08-14 anderson-cu-bioastronautics/dynamo_realsense-capture#6 @abhishektha
fixed by commit b972b86e91089126761db5548977be4226738cef
