### 2019-08-14 anderson-cu-bioastronautics/dynamo_realsense-capture#6 @alcantarar
Including this because it's partof the same [Camera Streaming to Disk Tutorial](https://github.com/anderson-cu-bioastronautics/dynamo_realsense-capture/blob/dynamo-release/examples/examples.md#camera-streaming-to-disk), but it has a typo: `import pyrealsense 2 as rs`. should be `import pyrealsense2 as rs`
