### 2019-08-14 anderson-cu-bioastronautics/dynamo_realsense-capture#6 @alcantarar
[Example Tutorial for Viewing Pointclouds](https://github.com/anderson-cu-bioastronautics/dynamo_realsense-capture/blob/dynamo-release/examples/examples.md#pointcloud-viewing-from-saved-frames) code has unused parameter `iteration` and misspelled function `viewPointClouds()`
