### 2019-08-14 anderson-cu-bioastronautics/dynamo_realsense-capture#9 @alcantarar
"RuntimeError: Camera already streaming" when trying to view pointclouds
When I run:
`view.viewPointClouds(folder,full=0)`

I get `RunTimeError: Camera already streaming` and print statements that say 2 devices have been found. The 3D interactive window pops up, but this RunTimeError is thrown repeatedly until I force stop the code. Ran in debug mode and it's breaking in `realsense_device_manager.enable_device()`.   Here's the full error:

```
2 devices have been found
Traceback (most recent call last):
  File "<string>", line 1, in <module>
  File "dynamo-env\lib\multiprocessing\spawn.py", line 105, in spawn_main
    exitcode = _main(fd)
  File "dynamo-env\lib\multiprocessing\spawn.py", line 114, in _main
    prepare(preparation_data)
  File "dynamo-env\lib\multiprocessing\spawn.py", line 225, in prepare
    _fixup_main_from_path(data['init_main_from_path'])
  File "dynamo-env\lib\multiprocessing\spawn.py", line 277, in _fixup_main_from_path
    run_name="__mp_main__")
  File "dynamo-env\lib\runpy.py", line 263, in run_path
    pkg_name=pkg_name, script_name=fname)
  File "dynamo-env\lib\runpy.py", line 96, in _run_module_code
    mod_name, mod_spec, pkg_name, script_name)
  File "dynamo-env\lib\runpy.py", line 85, in _run_code
    exec(code, run_globals)
  File "dynamo_realsense-capture\RA_Review.py", line 14, in <module>
    deviceManager.enable_all_devices()
  File "dynamo_realsense-capture\dynamo\realsense_device_manager.py", line 193, in enable_all_devices
    self.enable_device(serial, enable_ir_emitter)
  File "dynamo_realsense-capture\dynamo\realsense_device_manager.py", line 162, in enable_device
    pipeline_profile = pipeline.start(self._config)
RuntimeError: Camera already streaming
```

Again, the code keeps running but this error is thrown repeatedly.
