### 2019-08-14 bc6e97a
#6
removed unused parameter iteration in examples.md
