### 2019-08-14 06a139d
updated view.viewPointClouds() to better define default value for optional argument 'format'
