### 2019-08-14 anderson-cu-bioastronautics/dynamo_realsense-capture#9 @abhishektha
@alcantarar I'm unable to reproduce this. Are you trying to run `realsense_device_manager.enable_all_devices()` in a loop? This  error is usually seen when you are trying to enable the camera once it is already enabled. You should only need to enable the camera once per session.

I did just push an update (06a139dc55d2a85554b9abfe478c644ac4bc3b82) to `view.py` that should solve another bug I just found, don't think that would affect this issue but just wanted to let you know.
