### 2019-08-14 anderson-cu-bioastronautics/dynamo_realsense-capture#9 @alcantarar
@abhishektha I forked the repository and have added the pickled frames from `stream.start()` [here](https://github.com/alcantarar/dynamo_realsense-capture/tree/dynamo-release/TEST_DATA) and my implementation of  `view.viewPointClouds()` [here](https://github.com/alcantarar/dynamo_realsense-capture/blob/dynamo-release/RA_Review.py). That's all I can do right now to help with reproducing the error since I won't be back with the cameras until later this week.

(Part of https://github.com/openjournals/joss-reviews/issues/1466)
