### 2019-08-14 anderson-cu-bioastronautics/dynamo_realsense-capture#9 @abhishektha
Thanks for doing that, I'll investigate and get back to you ASAP!
