### 2019-08-28 anderson-cu-bioastronautics/dynamo_realsense-capture#9 @abhishektha
@alcantarar Still unable to recreate with your script nor files. If possible, would you mind testing if you can view your sample data with a new script where you simply try viewing the data?
