### 2019-09-04 #1466 @alcantarar
I've opened a few issues in the target repo that have been/are currently being addressed . I'll continue working on the [most recent issue](https://github.com/anderson-cu-bioastronautics/dynamo_realsense-capture/issues/9) with @abhishektha this week.
