### 2019-09-04 anderson-cu-bioastronautics/dynamo_realsense-capture#9 @alcantarar
@abhishektha I tried just running the following
```
view.viewPointClouds('C:/path/to/TEST_DATA', 0)
```
and get errors in `view.depthFrametoPC`:
```
    if not kwargs['format'] or kwargs['format'] == 'pcl':
KeyError: 'format'

 File "\dynamo_realsense-capture\dynamo\view.py", line 149, in viewPointClouds
    cameraPoints = p.map(depthFrametoPC, listP) #process point cloud for each camera in a separate thread
  File "\envs\dynamo-env\lib\multiprocessing\pool.py", line 266, in map
    return self._map_async(func, iterable, mapstar, chunksize).get()
  File "\envs\dynamo-env\lib\multiprocessing\pool.py", line 644, in get
    raise self._value
KeyError: 'format'
```
