### 2019-09-04 anderson-cu-bioastronautics/dynamo_realsense-capture#9 @abhishektha
@alcantarar, can you try the latest version of the code. commit 06a139dc55d2a85554b9abfe478c644ac4bc3b82 from a couple weeks ago should have addressed this problem and removed the reference to kwargs in the if statement
