### 2019-09-04 anderson-cu-bioastronautics/dynamo_realsense-capture#9 @alcantarar
Shoot, I missed that you'd made that commit. It works as intended now.
