### 2019-09-05 #1466 @alcantarar
hey @labarba my review is complete. There are a few open issues in the target repository about documentation, but @abhishektha can close them easily. I went all the way through the [Example Tutorial](https://github.com/anderson-cu-bioastronautics/dynamo_realsense-capture/blob/dynamo-release/examples/examples.md) and it functioned as advertised. Really fun using the cameras actually!

I've checked off the COI/COS and Functionality checklist above. Did I miss anything?
