### 2019-09-05 anderson-cu-bioastronautics/dynamo_realsense-capture#6 @abhishektha
closing after reviewing documentation
