### 2019-09-05 1a6f08c
#10
removed matplotlib import from dynamo.markers as it was never used and resulted in dependency errors
