+++ dynamo/markers.py
@@ -15,9 +15,7 @@ import sklearn
 from sklearn.cluster import DBSCAN
 import scipy.spatial
 
-from mpl_toolkits.mplot3d import Axes3D
 
-import matplotlib.pyplot as plt
 import os
 import sys
 
