### 2019-09-05 #1466 @abhishektha
@labarba, i went through and checked the documentation issues and addressed and closed them. Please let me know if there's anything else you need from me!
