### 2019-09-05 #1466 @labarba
@alcantarar — to clarify, you are only contributing a partial review, focusing on the functionality of the software, and do not plan to review the documentation and paper. Correct?
