### 2019-09-05 #1466 @alcantarar
@labarba Correct. I only made suggestions about documentation that directly influenced my ability to test the functionality. These issues have been addressed.
