### 2019-09-11 #1466 @alcantarar
The target repository claims to allow users to capture object shape or reflective marker data using multiple Intel RealSense cameras. The hardware requirements and software dependencies in the README.md file are comprehensive and allowed me to use my laptop to test two Intel RealSense cameras. The conda environment also worked as intended in terms of installation.

Setting up the cameras was straightforward and required no additional software to install. Users will be able to plug in the cameras and immediately be able to execute DynaMo functions to calibrate the cameras, stream/save images to their computer, and view 3D images afterwards. During my review, I ran into some issues with the number of parameters required for a given function or instances of hardcoding, but these issues have been addressed. Following the target repository's example page, I was able to record and save 3D object shape data from Intel RealSense Cameras, as the documentation claims.
