### 2019-09-26 45cd4d6
Minor edits
