+++ paper.md
@@ -22,19 +22,19 @@ bibliography: paper.bib
 
 # Background 
 
-Human body shape has been previously captured with a variety of methodologies, including laser lines, structured light, photogrammetry, and millimeter waves [@Dannen:2013].
-However these technologies require expensive modules and have limited ability to capture dynamic changes in body shape. 
+Human body shape can be captured with a variety of methodologies, including laser lines, structured light, photogrammetry, and millimeter waves [@Dannen:2013].
+However, these technologies require expensive modules and have limited ability to capture dynamic changes in body shape. 
 
-Similarly, motion capture with specific markers is commonly done through camera-based motion tracking [@Windolf:2008]
-Specific systems for marker tracking are also very cost prohibitive and do not capture surface morphology.
+Motion capture with specific markers is commonly done through camera-based motion tracking [@Windolf:2008]
+These systems for marker tracking are often cost prohibitive and unable to capture surface morphology.
 
-Recently, Intel has released the D415 and D435 RealSense Depth Cameras, which use near-infrared structured light patterns and two infrared imagers to capture depth information at up to 90 frames per second.
-Purchasing a set of these cameras is much more affordable than buying a dedicated motion capture system for shape or marker tracking.
+Recently, Intel released the D415 and D435 RealSense Depth Cameras, which use near-infrared structured light patterns and two infrared imagers to capture depth information at up to 90 frames per second.
+Purchasing a set of these cameras is more affordable than buying a dedicated motion-capture system for shape or marker tracking.
 
-While Intel provides the [``librealsense``](https://github.com/IntelRealSense/librealsense)  library to interface with their cameras, tools are not provided to use multiple devices at once to capture shape and marker-tracking information.
+While Intel provides the [``librealsense``](https://github.com/IntelRealSense/librealsense)  library to interface with their cameras, it lacks tools to use multiple devices at once to capture shape and marker-tracking information.
 ``DynaMo`` builds upon [``librealsense``](https://github.com/IntelRealSense/librealsense) to provide additional capability for researchers looking to capture such data.
 
-``DynaMo`` is designed to primarily assist those in the biomechanics and medical field in capturing motion capture or body shape data. 
+``DynaMo`` is designed to primarily assist those in the biomechanics and medical fields in capturing motion or body-shape data. 
 It is currently being used in the Anderson Bioastronautics Research Group to capture dynamic changes in foot morphology. 
 
 ![Sample Frames of DynaMo](documentation/sampleFrames.png)
@@ -42,9 +42,9 @@ It is currently being used in the Anderson Bioastronautics Research Group to cap
 
 # Summary
 
-``DynaMo`` is a Python library which provides tools to capture dynamic changes in body shape and track locations of markers using Intel RealSense D4XX cameras. 
+``DynaMo`` is a Python library that provides tools to capture dynamic changes in body shape and track locations of markers using Intel RealSense D4XX cameras. 
 ``DynaMo`` was developed from the examples provided by Intel in the Python [``librealsense``](https://github.com/IntelRealSense/librealsense) library. It has been successfully tested streaming six cameras at 90 frames per second, all connected to one computer. 
-``DynaMo`` consists of a number of scripts which allow for calibration of multiple RealSense D4XX cameras to a common global coordinate system, simultaneous streaming of multiple RealSense D4XX cameras, viewing of data from multiple RealSense D4XX cameras in pointcloud format, and identification of reflecting markers from the pointclouds.
[…]
