### 2019-09-26 #1466 @labarba
Please see my copy editing of the paper on the linked PR.
