### 2019-09-26 a6b92a7
fix missing DOIs
