+++ paper.bib
@@ -6,8 +6,8 @@ number = "4",
 pages = "270 - 275",
 year = "2013",
 issn = "0141-9382",
-doi = "https://doi.org/10.1016/j.displa.2013.08.011",
-url = "http://www.sciencedirect.com/science/article/pii/S014193821300070X",
+doi = {10.1016/j.displa.2013.08.011},
+url = "https://doi.org/10.1016/j.displa.2013.08.011",
 author = "H.A.M. Daanen and F.B. Ter Haar",
 keywords = "Whole-body-scanners, Anthropometry, Garments",
 abstract = "An overview of whole body scanners in 1998 (H.A.M. Daanen, G.J. Van De Water. Whole body scanners, Displays 19 (1998) 111–120) shortly after they emerged to the market revealed that the systems were bulky, slow, expensive and low in resolution. This update shows that new developments in sensing and processing technology, in particular in structured light scanners, have produced a new generation of easy to transport, fast, inexpensive, accurate and high resolution scanners. The systems are now moving to the consumer market with high impact for the garment industry. Since the internet sales of garments is rapidly increasing, information on body dimensions become essential to guarantee a good fit, and 3D scanners are expected to play a major role."
@@ -26,7 +26,8 @@ abstract={With rising demand on highly accurate acquisition of small motion the 
 keywords={Medical Sciences; Video-based motion analysis; Optical motion capturing; Accuracy; Precision; Vicon; Studies; Calibration; Robots; Cameras; Coordinate transformations},
 isbn={00219290},
 language={English},
-url={https://colorado.idm.oclc.org/login?url=https://search-proquest-com.colorado.idm.oclc.org/docview/1034929474?accountid=14503},
+doi = {10.1016/j.jbiomech.2008.06.024}
+url={https://doi.org/10.1016/j.jbiomech.2008.06.024},
 }
 
 @article{opencv_library,
