### 2019-09-26 61797fa
fixed missing comma in paper.bib
