+++ paper.bib
@@ -26,7 +26,7 @@ abstract={With rising demand on highly accurate acquisition of small motion the 
 keywords={Medical Sciences; Video-based motion analysis; Optical motion capturing; Accuracy; Precision; Vicon; Studies; Calibration; Robots; Cameras; Coordinate transformations},
 isbn={00219290},
 language={English},
-doi = {10.1016/j.jbiomech.2008.06.024}
+doi = {10.1016/j.jbiomech.2008.06.024},
 url={https://doi.org/10.1016/j.jbiomech.2008.06.024},
 }
 
