### 2019-09-26 b02cdf1
added missing reference to opencv library
