### 2019-09-26 #1466 @abhishektha
@labarba  I think the DOIs are okay now. I did fix a missing reference for the OpenCV library, which was not provided with a DOI.
