### 2019-09-29 #1466 @abhishektha
@labarba  All done:
tagged release: https://github.com/anderson-cu-bioastronautics/dynamo_realsense-capture/commit/b02cdf19bc0ffed766fe023ef05fe34dc5e29aba

zenodo DOI:
10.5281/zenodo.3464497

zenodo link: https://zenodo.org/record/3464497

Thanks!
