### 2019-09-29 #1466 @alcantarar
> Congratulations, @abhishekbajpayee, your JOSS paper is published! 🚀
>
> Thank you and hats off to our reviewers: @melund, @ixjlyons, @alcantarar — JOSS depends on you! 🙏

@labarba I think you mean "Congrats @abhishektha!"
